epochTimeToMSDOSDateTime: return minimum DOS datetime for earlier epoch times.
The previous behavior was just to crash.  Epoch times start in 1970, while
DOS datetimes start in 1980.  Returning the minimum seems the best solution,
since you simply can't create a zip archive entry with an earlier time.
This is also the solution used by zip30.

Thanks to Radoslav Dorcik for reporting.