Workaround for `toArchiveOrFail`.
It is impossible to write this function using the old binary
interface (< 0.7), which simply raises an error when parsing fails.
So, when compiled against binary < 0.7, this function will simply
call `toArchive` and slap on a Right; that is, it won't be safe.
This is advertised in the haddock comment.

This is ugly, but worth it to allow some form of the library to
be compiled against older versions of binary/ghc.

Closes #20.