toArchive eats too much memory when a data descriptor holds the size
The size fields in the local file headers may not contain valid values,
in which case the sizes are stored in a "data descriptor" that follows
the file data.

The handling of this case currently uses excessive amounts of memory,
because the entire compressed file is read at that point and temporarily
stored in a [Word8] list, which is rather wasteful for large streams.

Implement getWordsTilSig more efficiently by iteratively reading chunks
as strict ByteStrings and then converting these to a lazy ByteString at
the end.  This is a bit tricky because the magic separator 4 bytes long,
but only byte-aligned.