Test fixes.
The problem is that we can't include the test directory with
symlinks in extra-source-files, because this doesn't allow
a directory (or link to a directory) So, we have to build this
programatically.  This requires directory >= 1.3 in the test program.

I've added some tests, but they fail.  Not sure yet if this
is a problem in the test suite or the library itself. The error
is from getting the modification time of a symlink directory.
It may be that the library is not recognizing this as a directory.

Also fixed travis to build from an sdist, so we detect problems
like this in the future (files needed for test missing from
extra-source-files).