Remove existing symlinks when extracting zip files with symlinks (#60)
Previously, writeEntry would raise an error if it tried to create a symlink
and a symlink already existed at that path.  This behavior was inconsistent
with its behavior for regular files, which it overwrote without comment.
This commit causes symlinks to be replaced by writeEntry instead of
an error being raised.