Remove build-tools.
This was used to indicate that the 'unzip' executable was
needed for testing, but it was never intended to be used
this way and now the field is deprecated.

Supersedes #58.

The current test suite simply skips the test using the
unzip executable (with a warning) if 'unzip' is not in
the path.

See also #51 #41 #35 #23 #21.